Starting to tidy up the api package organisation, still work to do
